Add: Controller HistorialPrecios and finish Controller Ventas and their services

+++ Hache.Server/Entities/DetallePedido.cs
@@ -6,7 +6,7 @@
         public int ID_Pedido { get; set; }
         public int ID_Articulo { get; set; }
         public int Cantidad { get; set; }
-        public float Precio_Unitario { get; set; }
+        public decimal Precio_Unitario { get; set; }
         public DetallePedido(){ }
     }
 }
