+++ Hache.Server/Entities/HistorialPrecios.cs
@@ -4,8 +4,8 @@
     {
         public int ID_HistorialPrecios { get; set; }
         public int ID_Articulo { get; set; }
-        public float PrecioAnterior { get; set; }
-        public float PrecioNuevo { get; set; }
+        public decimal PrecioAnterior { get; set; }
+        public decimal PrecioNuevo { get; set; }
         public DateTime DateTime { get; set; } = DateTime.Now;
 
         public HistorialPrecios() { }
